fix(VideoPlayer): refactor ControlButtons and responsive video player
