+++ components/molecules/videoPlayer/helpers/videoSubtitle.js
@@ -1,0 +1,9 @@
+const videoSubtitle = ({ subtitles, subtitle }) => {
+  const track = (subtitles || []).find(({ srcLang = 'en' }) => srcLang === subtitle)
+  if (track) {
+    return <track {...track} default></track>
+  }
+  return null
+}
+
+export default videoSubtitle
